Isolate tests for hasMany.first
